refactor: suspense 사용 위해 수정하기 api 호출부를 template으로 내림
